refactor(validation): extract form schema projection
